+++ src/main/java/view/OutputView.java
@@ -1,9 +1,9 @@
 package view;
 
 import domain.Board;
-import domain.PieceColor;
 import domain.Player;
 import domain.piece.Piece;
+import domain.piece.PieceColor;
 import domain.piece.Pieces;
 import domain.piece.Position;
 import java.util.ArrayList;
